Add hook and update path
